+++ include/sora/open_h264_video_encoder.h
@@ -8,12 +8,17 @@
 #include <api/video_codecs/video_encoder.h>
 #include <media/base/codec.h>
 
+#include "sora_video_codec.h"
+
 namespace sora {
 
 std::unique_ptr<webrtc::VideoEncoder> CreateOpenH264VideoEncoder(
     const webrtc::SdpVideoFormat& format,
     std::string openh264);
 
-}
+VideoCodecCapability::Engine GetOpenH264VideoCodecCapability(
+    std::optional<std::string> openh264_path);
+
+}  // namespace sora
 
 #endif